Java Programming
Create a class Room which will store the length, width and height of the room in three private variables, constructor to initialize the variables, mutator and accessor methods, method to calculate the surface area, method to calculate the volume of the room. Create another Driver class which will use the earlier class, create instances of rooms, set the values of the variables and would calculate the surface area and volume of the room object.